+++ app/src/main/java/com/pict/pbl/medswift/screens/HomeScreen.kt
@@ -5,16 +5,17 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.viewModels
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.padding
+import androidx.compose.animation.core.SpringSpec
+import androidx.compose.animation.core.animateDpAsState
+import androidx.compose.animation.core.animateFloatAsState
+import androidx.compose.foundation.layout.*
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.*
 import androidx.compose.material3.*
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.draw.scale
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.platform.LocalConfiguration
@@ -97,6 +98,10 @@ class HomeScreen : ComponentActivity() {
             val navBackStackEntry by navController.currentBackStackEntryAsState()
             val currentRoute = navBackStackEntry?.destination?.route
             screens.forEach{ item ->
+                val top by animateFloatAsState(
+                    targetValue = if (currentRoute == item.screenRoute) 1.0f else 1.2f,
+                    animationSpec = SpringSpec(dampingRatio = 0.5f, stiffness = 200f)
+                )
                 NavigationBarItem(
                     selected = currentRoute == item.screenRoute,
                     onClick = {
@@ -110,7 +115,11 @@ class HomeScreen : ComponentActivity() {
                             restoreState = true
                         }
                     } ,
-                    icon = { Icon(imageVector = item.icon, contentDescription = item.title) } ,
+                    icon = {
+                        Icon(imageVector = item.icon,
+                            contentDescription = item.title ,
+                        modifier = Modifier.scale( top ) )
+                           } ,
                     label = { Text(text = item.title) }
                 )
             }